refactor: RestDocsUtils.prettyDocument 로 prettyPrint 생략
